revoke the special tag from ntcreatefile/ntopenfile
This has to do with the fact that when NtCreateFile/NtOpenFile are
called for every file operation we don't have enough available unicode
buffers. Thus some kind of reference counting has to be done in order
to ensure that we can work with a limited amount of buffers, or an
alternative solution has to be found.

This issue seems to be present when resolving the full filepath for a
filename. Especially around the code which calls GetFullPathNameW as
this seems to enumerate each underlying directory while calling
NtOpenFile on them, etc.