rewrite get-unicode-buffer so that the caller has to actually free it
This is required in order to handle calls that use more unicode buffers
than we support using a simple rotation algorithm. For example, if the
NtOpenFile function is defined as special signature, then a call to
GetFullPathNameW will call NtOpenFile one or more times on each part of
the directory given as path. As we would resolve each of these paths,
this would quickly result in going through the complete list of defined
unicode buffers, thus reusing buffers multiple times.