use callnamedpipew instead of callnamedpipe
In order not to clobber the per-thread unicode string we can just
provide a unicode string to begin with. This avoids having to back it
up or find alternative solutions, such as replacing the first argument
of NtDeleteFile by a copy of the original string after having called
pipe before-hand.