log exceptions right away
Now that the logging of exceptions is completely under our control we
can safely log exceptions as they come in, rather than having to wait
for another thread to pick up the exception, as that would previously
mess up the internal state of all kinds of things and crash even
harder. That is, we use our own sprintf implementation and we log the
bson data through a pipe with a native system call to NtWriteFile.