recursive api monitoring for dllmain
Due to the way the monitor works the monitor would not see calls when it
had already logged the "parent" call. In particular, when loading a DLL
into memory, at some point the Windows Loader will call DllMain of the
newly loaded DLL. Before any function calls in and around DllMain would
not be logged - now they will be.

This technique has not been tested yet against TLS callbacks but given
their similarity it would make sense if that works as well.