reuse pipe handler handle and go native
This change resolves various pipe related issues. Namely, before we
would use CallNamedPipe, but this would sometimes result in heap
allocation and other issues that caused for example the
NtAllocateVirtualMemory signature to crash under Internet Explorer 8 in
Windows 7. This work can be improved by further moving to native system
calls where this is not already the case.