improve inject.exe utility through the usage of ldrloaddll (thanks Accuvant)
Some processes may not have kernel32.dll loaded into the process and
thus we can't rely on LoadLibraryW being available. Brad Spengler
pointed out that this is the case for the critical system process
smss.exe.

This commit improves the way the injected code is represented. Instead
of manually assembling the injected code we use a function which takes
all its addresses from a structure provided upon injection. This
greatly simplifies modifying the injected code.

Also, instead of creating a new thread to close the handles obtained
through calling CreateProcessW, which is also a kernel32 function (but
we'll just assume that one will not use smss.exe or similar as starting
process), we use DuplicateHandle with the DUPLICATE_CLOSE_SOURCE flag
hoping that will suffice.