fix hooking issue in 64-bit mode with debugging enabled
In debug mode the monitor would be using a push/ret stub rather than a
32-bit relative jump. In 32-bit mode this is fine, however, in 64-bit
mode where most, if not all, of the loaded libraries reside in address
ranges not addressable by 32-bits this immediately gives us an exception
as execution flow reaches our first instance of a stub.

Note that this breaks all monitor unittests that require functionality
from native.c in 64-bit mode, but this can be fixed later. (It's not like
they would have worked anyway, though).