work in progress on logging all functions that are *not* hooked
This was a suggestion by Jop van der Lelie. Basically, it is possible
that a sample uses functions that we do not yet hook. In this case such
functions would go unhandled and we'd be missing out on potentially
interesting data.

There are still some recursion issues, but this commit shows work in
progress in this area. Later this should probably be configurable (and
disabled by default, probably).