handle delay-loaded function forwarders
In some situations an exported symbol will forward execution to another
DLL. If this other DLL is delay loaded then we can only hook the
function after the delay-loaded DLL has been loaded. Thanks to Accuvant
for first identifying this issue.

Instead of adding an additional hook for each function that is known to
be a delay-loaded function in some versions of Windows we dynamically
update the library of the hook to the delay-loaded library. This way we
automatically handle any new delay-loaded function forwarder hooks
transparently (or new versions of Windows that employ them for existing
hooks of ours).