dumping of tls keys
The introduced two hooks dump the Client Random, Server Random, and
Master Secret for each TLS/SSL3 handshake that is being performed. As
outlined in the following comment [1] these key computations are
performed for system-wide purposes in the lsass.exe system service. Thus
injecting our monitoring DLL (optionally just for extracting keys) into
lsass will yield us any keys used through the Windows TLS API (which
includes, but is probably not limited to, all wininet calls).

The work being performed by the monitor is only half of the total work.
As we're not able to extract the Session ID at this point (as far as I
can tell) we will do PCAP parsing on the host side to cross-reference
the found Client Random and/or Server Random between the instances and
finally link the Session ID (extracted from the PCAP) up to the Master
Secret found through the monitor.

This combination of information will allow, e.g., Wireshark to decrypt
all TLS traffic found in the PCAP file (as also outlined in [1]).

[1]: http://reverseengineering.stackexchange.com/a/1326