introduce the buffer message type
We have introduced a couple of functions for which it is necessary to
dump bigger buffers than the normal couple hundred bytes. We don't want
these big buffers directly embedded in the report though, so.. here
come the buffer files - dumped into the buffer/ directory of each
directory they're much like dropped files.