enable the now-working recursive dll loading technique
The previous commit fixed an issue which made debugging this feature
"suboptimal". However, now it's working perfectly fine. This technique
fixes the issue where LoadLibrary()'ing a DLL wouldn't hook the DLLs
imported by said DLL if they were not loaded already.

As an example, we have a COM DLL and analyze it with the COM package.
Now assume this COM DLL imports a function GetAdaptersInfo from a DLL
that is generally not loaded in processes by default, iphlpapi. Now as
iphlpapi would not be loaded through LoadLibrary/LdrLoadDll, but instead
through a lower level function, we would not hook its exported
functions, rendering our hooks useless.

This issue was also found and fixed by Optiv not too long ago.