unittests for imported dlls and hooked functions
The adapter unittest shows the GetAdaptersInfo() entry in the logs.
Before the changes to being able to hook recursively loaded DLLs, calls
to GetAdaptersInfo() and similar would not show up, rendering their
hooks useless.

The controlservice unittest is just to show that ControlService is being
hooked under 64-bit Windows 7, as it can't be hooked at once due to lazy
loading of DLLs (probably).