resolve recursive exceptions when logging exceptions
When logging regular function calls we also log the stacktrace in debug
mode. However, as we don't give a context to the stacktrace function,
and thus it tries to obtain the frame pointer from the ebp register, it
can't be used when logging exceptions as at the moment of
RtlDispatchException function call ebp is not correctly setup due to a
custom ABI being in use for KiUserExceptionDispatcher.