correct bounds checking through strlen-safe
We would provide a length of -1 for strings of which we didn't know the
actual length. This resulted in calling range_is_readable(ptr, ptr-1)
which defeats the point of checking boundaries in the first place.

Now we introduce a safe strlen and strlenW function that checks string
pointers before dereferencing them. Through these functions we can
improve the logging of strings and correctly perform boundary checks.