img.src assign hook
This can be used to extract Image URLs. Note that we're using a new
approach of locating the actual function here. Instead of trying to find
an overly complicated but reproducible way to locate a function among
multiple versions of a DLL we will simply keep track of their offsets.

It's not a future proof solution, but definitely the easiest to manage.
Naturally this is a first working version of the new approach. It should
be made more generic when we add other functions as well.