fix sign extension bugs in utf8 encoding (thanks knifeyspoony)
As pointed out in the comments it was possible for ascii or unicode
strings with the most significant bit set to effectively crash our
logging system as it'd be possible to create negative indexes into the
utf8 encoded string.

This commit fixes those issues and adds a bit of unittesting to make
sure this won't happen again in the future. This might fix a couple of
currently open bug reports [1] [2] [3].

[1]: https://github.com/cuckoosandbox/monitor/issues/21
[2]: https://github.com/cuckoosandbox/cuckoo/issues/841
[3]: https://github.com/cuckoosandbox/cuckoo/issues/853