replace function locating code by simply hardcoding addresses
In my opinion we can be better safe than sorry here. If an address was
resolved for a future version of, say, Internet Explorer, but the
address turns out to be incorrect, then we may have bigger problems than
a hook not executing, in the sense of debugging the issue (and
especially debugging the issue with users who report partial findings).

This commit thus restricts all IE8 related hooks to a specific version.
On request we can add hooks for different versions which should mostly
work correctly. Possibly we address issues from [1] [2] here.

[1]: https://github.com/cuckoosandbox/cuckoo/issues/841
[2]: https://github.com/cuckoosandbox/cuckoo/issues/853