always log createprocessinternalw and indicate if we track
In monitoring modes such as iexplore and office a new process may be
created without us being interested in that particular process. We
therefore ignore it (e.g., splwow64.exe for Office products) by not
injecting into it. By providing the track variable we can inform the
Host if it should account for said process or ignore it otherwise.