start on correctly handling invalid pointers
This commit radically changes the way that we handle pointers passed
along as parameters to invoked function calls. Rather than blindly
trusting that pointer parameters passed along actually point to the
information that we expect to see, we now take extra care when reading
such potentially invalid pointers.
This avoids issues where malware provides an invalid pointer that in
return crashes our monitor as it tries to dereference the pointer for
logging purposes.

The way that we have implemented this functionality is through a custom
version of a try/catch handler. Something that we already have available
since we hook at the lowest level of exception handling. From there on
we use setjmp(3)/longjmp(3) to return execution to the original code
path but then rather than copying again from an invalid pointer, return
either failure or simply none (i.e., 0 or a NULL pointer).