add wmi-based win32_process process following logic
By setting the CreateFlags flag of the Win32_ProcessStartup object that
is a parameter to the Win32_Process::Create function to CREATE_SUSPENDED
and manually figuring out the main thread identifier, it is possible for
us to follow process invocations using this WMI-based technique.

An example of such script in VBS may be as follows.

    set process = GetObject("winmgmts:Win32_Process")
    result = process.Create("notepad.exe", null, null, null)