dump flat ole1 streams for rtf documents
These OLE1 streams are basically deobfuscated RTF files, so - if parsed
correctly - they can be used to simplify the static analysis part of RTF
files. Based on the following FireEye blogpost [1].

On-the-go parsing of OLE2 streams (as OLE1 parsing has not been
implemented yet by any open source tool it appears) has yet to be done.

[1]: https://www.fireeye.com/blog/threat-research/2016/05/how_rtf_malware_evad.html