implement instruction-level hooking for version-specific hooks
This commit allows one to define instruction-level hooks (as in, not
necessarily tied to the start of a function) where desired to do so.
Two examples of instruction-level hooks are presented for specific
versions of Adobe 9 and Adobe 11.

The definition of this new type of hook is based on a YAML document that
specifies each hook - much like our custom RST-based document was doing,
but then in a slightly less custom way - one that with a bit of luck
will be used for instead of the RST-based documents at some point.

It should be noted that these instruction-level hooks are much like Brad
Spengler's "notail" hooks which operate on similar assumptions (e.g.,
don't leave return addresses from the Cuckoo Monitor in the stacktrace),
but that these instruction-level hooks may also be placed on
instructions which don't necessarily represent the function start.

If required one improvement would be to fetch an alternative "burner"
stack in which the Cuckoo Monitor related logging could be performed (so
not to overwrite memory at esp/rsp-0x1000).

Another possible application for the instruction-level hooks would to
build notail-like functionality to log function calls for which we are
also interested in capturing calls generated by the same call (e.g.,
LdrLoadDll will call DllMain of the target library).