simplistic unittest for reboot analysis
This test simply creates a registry key in HKCU\...\Run and exits. Then,
when restarted after reboot, it will take the first argument provided on
the command-line (which was set earlier in the registry key) and use
that to resolve a hostname.

By verifying that the host "evil.cuckoo.sh" is contacted during a reboot
analysis we may verify that the reboot analysis as well as this unittest
is functioning properly.