use guard pages to track arbitrary r/w exploitation payloads
Often times, an exploit achieves arbitrary read / write primitives
through one vulnerability or another. These primitives are then used to
scan memory regions for Windows DLLs, based on which an exploit / ROP
payload may be crafted. E.g., CVE-2015-2545 [1].

Using Guard Pages we're capable of intercepting these read operations on
Windows DLLs (and perhaps more later on?) so to identify that read /
write primitives are used in the first place, as well as potentially
identifying functions that are looked up by cross-referencing their
addresses.

This commit does not yet implement a notification method, whitelisting
of legitimate read operations (e.g., through GetProcAddress) or anything
of that matter, but that's yet to be implemented.

[1]: http://blog.morphisec.com/exploit-bypass-emet-cve-2015-2545