resolve alignment issues with setjmp on 64-bit architectures
Simply put mem_alloc() will allocate a memory page, put the length of
the allocation in the first uintptr-sized integer at the start of the
page, and return the rest of the memory page.
On 64-bit architectures this causes an issue as the pointer that will be
passed along to setjmp() therefore is not 16-byte aligned. This commit
fixes some issues including the misc.c unit test.