get rid of the messagebox to make inject.exe more silent
Thanks to danielgindi, rep, and most likely a whole bunch of others from
various Github issues for requesting this change. It doesn't make a
whole lot of sense for an automated system to wait for user input in
terms of clicking on a dialog box.

From now on any errors will simply go to stdout and the program will
exit with a failure (error code 1). This will allow the Cuckoo Analyzer
to throw around an exception including the error message (that was
printed to stdout in inject.exe) rather than waiting for the
non-existing user to click on something.

In future work any and all errors will be propagated to the various
Cuckoo reports and Cuckoo Web Interface. Although this wasn't possible
in the past (propagating the error towards the end user), it was
probably a bad decision to do it the way it has been up until now.