by popping into registers afterwards, we avoid write-before-read issues
In the case where a user would like to read one of the registers that
are used as part of the fastcall calling convention (i.e., rcx, rdx, r8,
and r9), but due to the ordering it's already overwritten, then we'll
find ourselves with some undefined behavior. This commit resolves that.